+++ butterfly.server.py
@@ -45,7 +45,8 @@ tornado.options.define("unminified", default=False,
 tornado.options.define("host", default='localhost', help="Server host")
 tornado.options.define("port", default=57575, type=int, help="Server port")
 tornado.options.define("keepalive_interval", default=30, type=int,
-                        help="Interval between ping packets sent from server to client (in seconds)")
+                       help="Interval between ping packets sent from server "
+                       "to client (in seconds)")
 tornado.options.define("one_shot", default=False,
                        help="Run a one-shot instance. Quit at term close")
 tornado.options.define("shell", help="Shell to execute at login")
@@ -54,10 +55,18 @@ tornado.options.define("cmd",
                        help="Command to run instead of shell, f.i.: 'ls -l'")
 tornado.options.define("unsecure", default=False,
                        help="Don't use ssl not recommended")
+tornado.options.define("i-hereby-declare-i-dont-want-any-security-whatsoever",
+                       default=False,
+                       help="Remove all security and warnings. There are some "
+                       "use cases for that. Use this if you really know what "
+                       "you are doing.")
 tornado.options.define("login", default=False,
                        help="Use login screen at start")
 tornado.options.define("pam_profile", default="", type=str,
-                       help="When --login=True provided and running as ROOT, use PAM with the specified PAM profile for authentication and then execute the user's default shell. Will override --shell.")
+                       help="When --login=True provided and running as ROOT, "
+                       "use PAM with the specified PAM profile for "
+                       "authentication and then execute the user's default "
+                       "shell. Will override --shell.")
 tornado.options.define("force_unicode_width",
                        default=False,
                        help="Force all unicode characters to the same width."
@@ -75,6 +84,7 @@ tornado.options.define("generate_user_pkcs", default='',
 tornado.options.define("uri_root_path", default='',
                        help="Sets the servier root path: "
                        "example.com/<uri_root_path>/static/")
+
 
 if os.getuid() == 0:
     ev = os.getenv('XDG_CONFIG_DIRS', '/etc')
@@ -130,6 +140,9 @@ log = logging.getLogger('butterfly')
 
 host = options.host
 port = options.port
+
+if options.i_hereby_declare_i_dont_want_any_security_whatsoever:
+    options.unsecure = True
 
 
 if not os.path.exists(options.ssl_dir):