+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+3.1.4
+=====
+
+* Add --i-hereby-declare-i-dont-want-any-security-whatsoever option (#143)
+
+3.1.3
+=====
+
+* Fix lsof parsing crash on python 2
+
 3.1.0
 =====